회원 쿠폰 클래스 추출
